Move file tailer's openFile to pkg/util/filesystem
This is a windows-only reimplementation of os.Open that allows
(some) concurrent use of the file.  It will be useful in checking
file access in the container launcher.
