+++ pkg/logs/internal/tailers/file/tailer_nix.go
@@ -13,6 +13,7 @@ import (
 	"path/filepath"
 
 	"github.com/DataDog/datadog-agent/pkg/logs/internal/decoder"
+	"github.com/DataDog/datadog-agent/pkg/util/filesystem"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
 
@@ -28,7 +29,7 @@ func (t *Tailer) setup(offset int64, whence int) error {
 	t.tags = t.buildTailerTags()
 
 	log.Info("Opening", t.file.Path, "for tailer key", t.file.GetScanKey())
-	f, err := openFile(fullpath)
+	f, err := filesystem.OpenShared(fullpath)
 	if err != nil {
 		return err
 	}
